// packages/radio/types/iradio.ts
interface IRadio {
  ReactThemeContext?: any;
  backgroundColor?: string;
  checked?: boolean;
  color?: string;
  disabled?: boolean;
  fontFamily?: string;
  fontSize?: string;
  id?: string;
  label?: string;
  name?: string;
  onChange?: (evt: any) => void;
  theme?: any;
}

export default IRadio;